+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/Timeline.kt
@@ -1,5 +1,7 @@
 package de.connect2x.messenger.compose.view.room.timeline
 
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -266,8 +268,10 @@ class TimelineViewImpl : TimelineView {
                                 contentPadding = PaddingValues(
                                     top = 10.dp,
                                     bottom = 10.dp,
-                                    start = if (this@BoxWithConstraints.maxWidth.value > 1000) 80.dp else 10.dp,
-                                    end = if (this@BoxWithConstraints.maxWidth.value > 1000) 80.dp else 18.dp, // 10 + 8, since we cannot add a padding or Spacer at the end
+                                    start = if (this@BoxWithConstraints.maxWidth.value > 1000)
+                                            (0.5 * (this@BoxWithConstraints.maxWidth.value - 1000) + 10).dp else 10.dp,
+                                    end = if (this@BoxWithConstraints.maxWidth.value > 1000)
+                                            (0.5 * (this@BoxWithConstraints.maxWidth.value - 1000) + 18).dp else 18.dp, // 10 + 8, since we cannot add a padding or Spacer at the end
                                 ),
                                 state = listState,
                                 reverseLayout = true,