+++ CHANGELOG.md
@@ -16,10 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored message bubbles to support message previews
 - Refactored read receipts and reactions to be more modular
+- Insert formula in padding of the timeline
 
 ### Deprecated
 
 ### Removed
+
+- Remove padding applied to app layout for sizes above a certain threshold
 
 ### Fixed